check to see that the file exists before deleting it.  This is a new bug that was introduced as we started checking return from delete.  I didn't realize delete was often used without the file existing, which worked because we ignored the status.
git-svn-id: https://svn.code.sf.net/p/autoplot/code/autoplot/trunk@8348 0c594bb0-8543-0410-a25e-d8ec440261db

+++ VirboAutoplot/src/com/cottagesystems/jdiskhog/JDiskHogPanel.java
@@ -63,7 +63,9 @@ public class JDiskHogPanel extends javax.swing.JPanel {
                         continue;
                     }
                     if (f.isFile()) {
-                        okay = f.delete();
+                        if ( f.exists() ) {
+                            okay = f.delete();
+                        }
                     } else {
                         try {
                             okay = Util.deleteFileTree(f);
